# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for CerdasIND Examination System",
        "title": "CerdasIND API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticate user and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new participant account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register participant",
                "parameters": [
                    {
                        "description": "Register Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bundles/{id}/review": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve questions with keys and discussion after exam is finished",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Get exam review",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bundle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Soal"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bundles/{id}/soal": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve questions for an exam bundle (hides keys and discussion)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Get soal by bundle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bundle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SoalPublic"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bundles/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit answers for an exam bundle and get auto-grade result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Submit exam answers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bundle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submit Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SubmitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jenjang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve list of all educational levels (SD, SMP, SMA)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Get all jenjang",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Jenjang"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jenjang/{id}/mapel": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve list of subjects based on jenjang ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Get mapel by jenjang",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jenjang ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Mapel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mapel/{id}/bundles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve list of active exam bundles based on mapel ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Get bundles by mapel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mapel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Bundle"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve exam history for the logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participant"
                ],
                "summary": "Get user history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.HistoryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AuthResponse": {
            "type": "object",
            "properties": {
                "role": {
                    "$ref": "#/definitions/model.UserRole"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.Bundle": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deskripsi": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "mapel_id": {
                    "type": "integer"
                },
                "nama_bundle": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "waktu_menit": {
                    "type": "integer"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.HistoryResponse": {
            "type": "object",
            "properties": {
                "history_id": {
                    "type": "integer"
                },
                "nama_bundle": {
                    "type": "string"
                },
                "skor_akhir": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/model.StatusUjian"
                },
                "waktu_mulai": {
                    "type": "string"
                }
            }
        },
        "model.JenisSoal": {
            "type": "string",
            "enum": [
                "pilihan_ganda",
                "isian_singkat"
            ],
            "x-enum-varnames": [
                "SoalPG",
                "SoalIsianSingkat"
            ]
        },
        "model.Jenjang": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
        "model.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.Mapel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jenjang_id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "model.PilihanJawaban": {
            "type": "object",
            "properties": {
                "opsi": {
                    "type": "string"
                },
                "teks": {
                    "type": "string"
                }
            }
        },
        "model.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.Soal": {
            "type": "object",
            "properties": {
                "bobot_nilai": {
                    "type": "integer"
                },
                "bundle_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kunci_jawaban": {
                    "type": "string"
                },
                "pembahasan": {
                    "type": "string"
                },
                "pilihan_jawaban": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PilihanJawaban"
                    }
                },
                "teks_soal": {
                    "type": "string"
                },
                "tipe_soal": {
                    "$ref": "#/definitions/model.JenisSoal"
                }
            }
        },
        "model.SoalPublic": {
            "type": "object",
            "properties": {
                "bobot_nilai": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "pilihan_jawaban": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PilihanJawaban"
                    }
                },
                "teks_soal": {
                    "type": "string"
                },
                "tipe_soal": {
                    "$ref": "#/definitions/model.JenisSoal"
                }
            }
        },
        "model.StatusUjian": {
            "type": "string",
            "enum": [
                "berlangsung",
                "menunggu_koreksi",
                "selesai"
            ],
            "x-enum-varnames": [
                "StatusBerlangsung",
                "StatusMenungguKoreksi",
                "StatusSelesai"
            ]
        },
        "model.SubmitJawaban": {
            "type": "object",
            "properties": {
                "jawaban_peserta": {
                    "type": "string"
                },
                "soal_id": {
                    "type": "integer"
                }
            }
        },
        "model.SubmitRequest": {
            "type": "object",
            "properties": {
                "jawaban": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubmitJawaban"
                    }
                }
            }
        },
        "model.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "peserta"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RolePeserta"
            ]
        },
        "model.WebResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}