# internal/model/user.go
package model

import "time"

type UserRole string

const (
	RoleAdmin   UserRole = "admin"
	RolePeserta UserRole = "peserta"
)

type User struct {
	ID           int64     `json:"id" db:"id"`
	Username     string    `json:"username" db:"username"`
	Email        string    `json:"email" db:"email"`
	PasswordHash string    `json:"-" db:"password_hash"`
	Role         UserRole  `json:"role" db:"role"`
	CreatedAt    time.Time `json:"created_at" db:"created_at"`
}

type LoginRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required"`
}

type RegisterRequest struct {
	Username string `json:"username" binding:"required"`
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6"`
}

type AuthResponse struct {
	UserID   int64    `json:"user_id"`
	Username string   `json:"username"`
	Role     UserRole `json:"role"`
	Token    string   `json:"token"`
}